# SP ECC Cache Resitance