… wolfcrypt/src/sp_arm32.c
… wolfcrypt/src/sp_arm64.c
… wolfcrypt/src/sp_armthumb.c
… wolfcrypt/src/sp_c32.c
… wolfcrypt/src/sp_c64.c
… wolfcrypt/src/sp_cortexm.c
… wolfcrypt/src/sp_x86_64.c