@toddouska 2020-08-06
What's the performance delta with touching every point before every use?