@SparkiDev 2020-08-06
Summary:
~17% slower on key gen
~5% slower on key agree
~12% slower on sign
No effect on verify

Still faster than OpenSSL.

Customer's should turn on WC_NO_CACHE_RESISTANT if they now they don't need it.


Master:
------------------------------------------------------------------------------
 wolfSSL version 4.4.1
------------------------------------------------------------------------------
wolfCrypt Benchmark (block bytes 1048576, min 1.0 sec each)
ECC      256 key gen    114807 ops took 1.000 sec, avg 0.009 ms, 114806.316 ops/sec
ECDHE    256 agree       29100 ops took 1.001 sec, avg 0.034 ms, 29059.407 ops/sec
ECDSA    256 sign        71600 ops took 1.001 sec, avg 0.014 ms, 71539.774 ops/sec
ECDSA    256 verify      23500 ops took 1.001 sec, avg 0.043 ms, 23482.739 ops/sec
Benchmark complete

New code:
------------------------------------------------------------------------------
 wolfSSL version 4.4.1
------------------------------------------------------------------------------
wolfCrypt Benchmark (block bytes 1048576, min 1.0 sec each)
ECC      256 key gen     95635 ops took 1.000 sec, avg 0.010 ms, 95634.339 ops/sec
ECDHE    256 agree       28100 ops took 1.004 sec, avg 0.036 ms, 28001.322 ops/sec
ECDSA    256 sign        63300 ops took 1.000 sec, avg 0.016 ms, 63296.710 ops/sec
ECDSA    256 verify      23600 ops took 1.001 sec, avg 0.042 ms, 23568.325 ops/sec
Benchmark complete

OpenSSL 3.0.0-dev:
                              sign    verify    sign/s verify/s
 256 bits ecdsa (nistp256)   0.0000s   0.0000s  58592.3  20461.2
                              op      op/s
 256 bits ecdh (nistp256)   0.0000s  27005.7

OpenSSL 1.1.1f  31 Mar 2020
                              sign    verify    sign/s verify/s
 256 bits ecdsa (nistp256)   0.0000s   0.0000s  59671.6  20235.8
                              op      op/s
 256 bits ecdh (nistp256)   0.0000s  26971.0